Refactor SMART discovery and metadata handling; remove obsolete classes and implement new structure
