+++ lib/safire.rb
@@ -1,8 +1,15 @@
 require 'active_support/all'
+
+root = File.expand_path '.', File.dirname(File.absolute_path(__FILE__))
+Dir.glob(File.join(root, 'safire', 'client', '**', '*.rb')).each do |file|
+  require file
+end
+
 require_relative 'safire/version'
 require_relative 'safire/errors'
 require_relative 'safire/http_client'
-require_relative 'safire/logger'
+require_relative 'safire/safire_logger'
+require_relative 'safire/safire_client'
 
 # Main module for Safire gem
 module Safire
@@ -19,7 +26,7 @@ module Safire
     attr_accessor :logger, :log_level, :timeout, :user_agent
 
     def initialize
-      @logger = Safire::Logger.new
+      @logger = Safire::SafireLogger.new
       @log_level = :info
       @timeout = 30
       @user_agent = "Safire v#{Safire::VERSION}"
