--- lib/safire/client/entity.rb
@@ -1,9 +1,0 @@
-module Safire
-  module Client
-    class Entity
-      def initialize(params, attributes)
-        attributes.each { |name| instance_variable_set("@#{name}", params[name] || params[name.to_s]) }
-      end
-    end
-  end
-end
