+++ lib/safire/entity.rb
@@ -1,0 +1,7 @@
+module Safire
+  class Entity
+    def initialize(params, attributes)
+      attributes.each { |name| instance_variable_set("@#{name}", params[name] || params[name.to_s]) }
+    end
+  end
+end
